Thêm UpLoad ảnh vào addService

+++ database/migrations/2025_04_15_023545_create_employee_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('full_name', 255);
             $table->string('position', 255);
-            $table->decimal('wage', 10, 2);
+            $table->integer('wage');
             $table->string('phone', 15)->unique();
             $table->string('email', 255)->unique();
             $table->timestamps();
